requirements.txt
streamlit
python-dotenv
pandas
pdfplumber
google-generativeai
openpyxl
